--- components/examine/RequestHeader.vue
@@ -1,43 +1,0 @@
-<template>
-  <div class="flex flex-wrap px-4 py-2 text-gray-700">
-    <ExamineHeaderNRNumber
-      :nr-number="examine.nrNumber"
-      :priority="examine.isPriority"
-      show-transactions-link
-    />
-
-    <ExamineHeaderRequestTypeInfo
-      v-if="!examine.is_editing || examine.isClosed"
-      :jurisdiction="examine.jurisdiction"
-      :jurisdiction-number="examine.jurisdictionNumber"
-    />
-    <ExamineHeaderRequestEditInputs v-else class="flex basis-5/12 flex-col" />
-
-    <ExamineActionButtons class="ml-auto" />
-  </div>
-</template>
-
-<script setup lang="ts">
-import {
-  EntityTypeCode,
-  RequestActionCode,
-  RequestTypeCode,
-} from '~/enums/codes'
-import { useExamineStore } from '~/store/examine'
-import { getMappedRequestType } from '~/util'
-
-const examine = useExamineStore()
-
-const isViewing = computed(() => examine.is_header_shown && !examine.is_editing)
-const isExpanded = computed(
-  () => examine.is_editing || isViewing || examine.is_header_shown
-)
-
-const requestType = computed(() =>
-  getMappedRequestType(
-    examine.requestType as RequestTypeCode,
-    examine.requestActionCd as RequestActionCode,
-    examine.entityTypeCd as EntityTypeCode
-  )
-)
-</script>
